+++ TestApp/include/perf/Tests/EstimatorPerfTests.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Performance/PerfTest.hpp"
+#include "perf/PerfTest.hpp"
 #include "Precision.hpp"
 
 class LinearSplinePerfTest : public TimedTest
@@ -18,7 +18,7 @@ private:
 public:
 	LinearSplinePerfTest(
 		std::string&& name,
-		const std::variant<uint64_t, std::chrono::nanoseconds> limit,
+		const std::variant<size_t, std::chrono::nanoseconds> limit,
 		float_s(*generator) (const float_s),
 		const uint32_t size,
 		const float_s loss
