+++ TestApp/src/Unit/Tests/TimingUnitTests.cpp
@@ -1,11 +1,11 @@
-#include "Unit/Tests/TimingUnitTests.hpp"
-#include "Performance/PerformanceReport.hpp"
+#include "unit/tests/TimingUnitTests.hpp"
+#include "perf/PerformanceReport.hpp"
 #include "Log.hpp"
 
 #include <thread>
 
 TimingUnitTest::DummyPerfTest::DummyPerfTest(
-	const std::variant<uint64_t, std::chrono::nanoseconds> limit,
+	const std::variant<size_t, std::chrono::nanoseconds> limit,
 	const std::chrono::nanoseconds waitTime
 ) noexcept :
 	TimedTest("dummy", limit),
@@ -26,12 +26,12 @@ std::chrono::nanoseconds TimingUnitTest::DummyPerfTest::getWaitTime() const
 
 TimingUnitTest::TimingUnitTest(
 	const std::string& name,
-	const std::variant<uint64_t, std::chrono::nanoseconds> limit,
+	const std::variant<size_t, std::chrono::nanoseconds> limit,
 	const std::chrono::nanoseconds waitTime,
 	const float_h threshold
 ) noexcept :
-	UnitTest(name + '_' + (std::holds_alternative<uint64_t>(limit) ?
-		std::to_string(std::get<uint64_t>(limit)) :
+	UnitTest(name + '_' + (std::holds_alternative<size_t>(limit) ?
+		std::to_string(std::get<size_t>(limit)) :
 		std::to_string(std::get<std::chrono::nanoseconds>(limit).count()) + "ns") +
 		'_' + std::to_string(waitTime.count()) + "ns"),
 	perfTest(limit, waitTime),
@@ -63,13 +63,13 @@ TimingUnitTests::TimingUnitTests(
 ) noexcept :
 	UnitTestGroup(std::move(name), filter)
 {
-	registerTest<TimingUnitTest>("counted", uint64_t(20), std::chrono::milliseconds(10), 0.3);
-	registerTest<TimingUnitTest>("counted", uint64_t(100), std::chrono::milliseconds(10), 0.2);
+	registerTest<TimingUnitTest>("counted", size_t(20), std::chrono::milliseconds(10), 0.3);
+	registerTest<TimingUnitTest>("counted", size_t(100), std::chrono::milliseconds(10), 0.2);
 
-	registerTest<TimingUnitTest>("counted", uint64_t(30), std::chrono::microseconds(100), 5.0);
-	registerTest<TimingUnitTest>("counted", uint64_t(20), std::chrono::milliseconds(1), 3.0);
-	registerTest<TimingUnitTest>("counted", uint64_t(20), std::chrono::milliseconds(10), 0.1);
-	registerTest<TimingUnitTest>("counted", uint64_t(10), std::chrono::milliseconds(100), 0.06);
+	registerTest<TimingUnitTest>("counted", size_t(30), std::chrono::microseconds(100), 5.0);
+	registerTest<TimingUnitTest>("counted", size_t(20), std::chrono::milliseconds(1), 3.0);
+	registerTest<TimingUnitTest>("counted", size_t(20), std::chrono::milliseconds(10), 0.1);
+	registerTest<TimingUnitTest>("counted", size_t(10), std::chrono::milliseconds(100), 0.06);
 
 	registerTest<TimingUnitTest>("timed", std::chrono::milliseconds(200), std::chrono::milliseconds(10), 0.3);
 	registerTest<TimingUnitTest>("timed", std::chrono::seconds(1), std::chrono::milliseconds(10), 0.2);
